+++ crates/client/rpc/src/runtime_api.rs
@@ -11,15 +11,12 @@ pub use mc_rpc_core::{
 use mp_felt::Felt252Wrapper;
 use mp_hashers::HasherT;
 use mp_simulations::SimulationFlags;
-use pallet_starknet_runtime_api::{
-    ConvertTransactionRuntimeApi, StarknetRuntimeApi, StarknetTransactionExecutionError,
-};
+use pallet_starknet_runtime_api::{ConvertTransactionRuntimeApi, StarknetRuntimeApi};
 use sc_client_api::backend::Backend;
 use sc_transaction_pool::ChainApi;
-use sp_api::ProvideRuntimeApi;
+use sp_api::{ApiError, ProvideRuntimeApi};
 use sp_blockchain::HeaderBackend;
 use sp_runtime::traits::Block as BlockT;
-use sp_runtime::DispatchError;
 use starknet_api::core::{ContractAddress, EntryPointSelector};
 use starknet_api::transaction::{Calldata, Event, TransactionHash};
 
@@ -43,11 +40,13 @@ where
         contract_address: ContractAddress,
         entry_point_selector: EntryPointSelector,
         calldata: Calldata,
-    ) -> RpcApiResult<Result<Vec<Felt252Wrapper>, sp_runtime::DispatchError>> {
-        self.client.runtime_api().call(best_block_hash, contract_address, entry_point_selector, calldata).map_err(|e| {
-            error!("Request parameters error: {e}");
-            StarknetRpcApiError::InternalServerError
-        })
+    ) -> RpcApiResult<Vec<Felt252Wrapper>> {
+        Ok(self.client.runtime_api().call(best_block_hash, contract_address, entry_point_selector, calldata).map_err(
+            |e| {
+                error!("Request parameters error: {e}");
+                StarknetRpcApiError::InternalServerError
+            },
+        )??)
     }
 
     pub fn do_estimate_message_fee(
@@ -55,17 +54,10 @@ where
         block_hash: B::Hash,
         message: L1HandlerTransaction,
     ) -> RpcApiResult<(u128, u128, u128)> {
-        self.client
-            .runtime_api()
-            .estimate_message_fee(block_hash, message)
-            .map_err(|e| {
-                error!("Runtime Api error: {e}");
-                StarknetRpcApiError::InternalServerError
-            })?
-            .map_err(|e| {
-                error!("Function execution failed: {:#?}", e);
-                StarknetRpcApiError::ContractError
-            })
+        Ok(self.client.runtime_api().estimate_message_fee(block_hash, message).map_err(|e| {
+            error!("Runtime Api error: {e}");
+            StarknetRpcApiError::InternalServerError
+        })???)
     }
 
     pub fn do_get_tx_execution_outcome(
@@ -96,17 +88,6 @@ where
         })
     }
 
-    pub fn convert_dispatch_error(
-        &self,
-        best_block_hash: B::Hash,
-        error: DispatchError,
-    ) -> RpcApiResult<StarknetTransactionExecutionError> {
-        self.client.runtime_api().convert_error(best_block_hash, error).map_err(|e| {
-            error!("Failed to convert dispatch error: {:?}", e);
-            StarknetRpcApiError::InternalServerError
-        })
-    }
-
     pub fn convert_tx_to_extrinsic(
         &self,
         best_block_hash: <B as BlockT>::Hash,
@@ -124,17 +105,12 @@ where
         transactions: Vec<AccountTransaction>,
         simulation_flags: SimulationFlags,
     ) -> RpcApiResult<Vec<(u128, u128)>> {
-        self.client
-            .runtime_api()
-            .estimate_fee(block_hash, transactions, simulation_flags)
-            .map_err(|e| {
+        Ok(self.client.runtime_api().estimate_fee(block_hash, transactions, simulation_flags).map_err(
+            |e: ApiError| {
                 error!("Request parameters error: {e}");
                 StarknetRpcApiError::InternalServerError
-            })?
-            .map_err(|e| {
-                error!("Failed to call function: {:#?}", e);
-                StarknetRpcApiError::ContractError
-            })
+            },
+        )???)
     }
 
     pub fn get_best_block_hash(&self) -> B::Hash {
@@ -206,8 +182,8 @@ where
             })?
             .map_err(|e| {
                 error!("Failed to call function: {:#?}", e);
-                StarknetRpcApiError::ContractError
-            })?
+                StarknetRpcApiError::from(e)
+            })??
             .swap_remove(0)
             .1
             .map_err(|e| {
@@ -232,7 +208,7 @@ where
             })?
             .map_err(|e| {
                 error!("Failed to call function: {:#?}", e);
-                StarknetRpcApiError::ContractError
+                StarknetRpcApiError::from(e)
             })?
             .map_err(|e| {
                 error!("Failed to simulate L1 Message: {:?}", e);